rename pattern repository to patterns repository
