+++ lib/patterns_repository/core/cartesian_product_iterable.dart
