+++ lib/patterns_repository/core/compare.dart
