+++ lib/patterns_repository/core/config_map.dart
