+++ lib/patterns_repository/core/core.dart
