+++ lib/patterns_repository/core/rotateable_list.dart
