+++ lib/patterns_repository/engine/prechac_pattern.dart
