+++ lib/patterns_repository/engine/prechac_pattern_constraint.dart
