+++ lib/patterns_repository/engine/prechac_throw_constraint.dart
