+++ lib/patterns_repository/models/models.dart
