+++ lib/patterns_repository/models/pattern.dart
