+++ lib/patterns_repository/models/pattern_constraint.dart
