+++ lib/patterns_repository/models/patternable.dart
