+++ lib/patterns_repository/models/search_parameters.dart
