+++ lib/patterns_repository/models/throw.dart
