+++ lib/patterns_repository/models/throw_constraint.dart
