+++ lib/patterns_repository/models/throwable.dart
