+++ lib/screens/search_results/view/search_results_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../pattern_repository/pattern_repository.dart';
+import '../../../patterns_repository/patterns_repository.dart';
 import '../search_results.dart';
 
 class SearchResultsPage extends StatelessWidget {
@@ -24,7 +24,7 @@ class SearchResultsPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return BlocProvider(
       create: (_) => PatternsBloc(
-        patternRepository: PatternRepository(),
+        patternsRepository: PatternsRepository(),
       )..add(LoadPatterns(searchParameters)),
       child: SearchResultsView(),
     );
