+++ lib/screens/search_results/widgets/pattern_item.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import '../../../pattern_repository/pattern_repository.dart';
+import '../../../patterns_repository/patterns_repository.dart';
 
 class PatternItem extends StatelessWidget {
   PatternItem({
