+++ test/patterns_repository/engine/prechac_pattern_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:prechac_this/pattern_repository/engine/prechac_pattern.dart';
+import 'package:prechac_this/patterns_repository/engine/prechac_pattern.dart';
 
 import '../../helpers/helpers.dart';
