+++ test/screens/search_results/bloc/patterns_bloc_test.dart
@@ -2,24 +2,24 @@
 import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:prechac_this/pattern_repository/pattern_repository.dart';
+import 'package:prechac_this/patterns_repository/patterns_repository.dart';
 import 'package:prechac_this/screens/search_results/search_results.dart';
 
 import '../../../helpers/helpers.dart';
 
-class MockPatternRepository extends Mock implements PatternRepository {}
+class MockPatternsRepository extends Mock implements PatternsRepository {}
 
 void main() {
   group('PatternsBloc', () {
-    late PatternRepository patternRepository;
+    late PatternsRepository patternsRepository;
 
     setUp(() {
-      patternRepository = MockPatternRepository();
+      patternsRepository = MockPatternsRepository();
     });
 
     test('initial state is HomeState', () {
       expect(
-        PatternsBloc(patternRepository: patternRepository).state,
+        PatternsBloc(patternsRepository: patternsRepository).state,
         PatternsInitial(),
       );
     });
@@ -27,9 +27,9 @@ void main() {
     blocTest<PatternsBloc, PatternsState>(
       'emits [PatternsLoading, PatternsLoaded]',
       build: () {
-        when(() => patternRepository.patterns(mockParameters))
+        when(() => patternsRepository.patterns(mockParameters))
             .thenAnswer((invocation) async => [mockPattern]);
-        return PatternsBloc(patternRepository: patternRepository);
+        return PatternsBloc(patternsRepository: patternsRepository);
       },
       act: (bloc) => bloc.add(LoadPatterns(mockParameters)),
       expect: () => <PatternsState>[
